Added commands which reference the database. Now predicates uses the Builder pattern.

+++ Models/RequestModels.cs
@@ -1,11 +1,11 @@
 namespace Backend.Models {
     public class TwoIds {
-        public int WinnerId { get; set; }
-        public int LoserId { get; set; }
+        public ulong WinnerId { get; set; }
+        public ulong LoserId { get; set; }
         public string Topic { get; set; }
     }
 
     public class OneId {
-        public int PlayerId { get; set; }
+        public ulong PlayerId { get; set; }
     }
 }
